while i would have liked to use MSAA, using MRT is just to useful.  sometime in the future i'll have to implement FXAA.  added id, color, glow, and soon depth render targets.

+++ main/lucid/gui/Button.h
@@ -27,7 +27,7 @@ public:
 
 	typedef std::array<LUCID_GAL::Vector4, STATE_COUNT> Tiles;
 
-	Button(size_t id, ANCHOR anchor, Size const &size, Tiles const &tiles, Callback const &callback);
+	Button(uint32_t id, ANCHOR anchor, Size const &size, Tiles const &tiles, Callback const &callback);
 
 	virtual ~Button() = default;
