+++ main/lucid/gui/Label.cpp
@@ -3,7 +3,7 @@
 
 LUCID_GUI_BEGIN
 
-Label::Label(size_t id, ANCHOR anchor, Size const &size, ALIGNMENT align, std::string const &text, Color const &color)
+Label::Label(uint32_t id, ANCHOR anchor, Size const &size, ALIGNMENT align, std::string const &text, Color const &color)
 	: Frame(id, anchor, size)
 	, _alignment(align)
 	, _text(text)
